Feature/dental clinic/add service layer

+++ src/main/java/com/dh/Dental/Clinic/service/impl/DentistService.java
@@ -1,5 +1,6 @@
 package com.dh.Dental.Clinic.service.impl;
 
+import com.dh.Dental.Clinic.dto.response.DentistResponseDto;
 import com.dh.Dental.Clinic.entity.Dentist;
 
 import com.dh.Dental.Clinic.entity.Patient;
@@ -9,6 +10,7 @@ import com.dh.Dental.Clinic.service.IDentistService;
 
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import java.util.Optional;
@@ -28,22 +30,45 @@ public class DentistService implements IDentistService {
     }
 
     @Override
-    public Optional<Dentist> findDentistById(Integer id) {
-        return dentistRepository.findById(id);
+    public Optional<DentistResponseDto> findDentistById(Integer id) {
+        Optional<Dentist> dentist = dentistRepository.findById(id);
+        return dentist.map(this::convertDentistToResponse);
     }
 
     @Override
-    public List<Dentist> findAllDentists() {
-        return dentistRepository.findAll();
+    public List<DentistResponseDto> findAllDentists() {
+        List<Dentist> dentists = dentistRepository.findAll();
+        List<DentistResponseDto> responseDtos = new ArrayList<>();
+        for (Dentist dentist : dentists) {
+            responseDtos.add(convertDentistToResponse(dentist));
+        }
+        return responseDtos;
     }
 
     @Override
     public void updateDentist(Dentist dentist) {
-        dentistRepository.save(dentist);
+        if (dentistRepository.existsById(dentist.getId())) {
+            dentistRepository.save(dentist);
+        } else {
+            throw new RuntimeException("Dentist not found");
+        }
     }
 
     @Override
     public void deleteDentist(Integer id) {
-        dentistRepository.deleteById(id);
+        if (dentistRepository.existsById(id)) {
+            dentistRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Dentist not found");
+        }
+    }
+
+    private DentistResponseDto convertDentistToResponse(Dentist dentist) {
+        return new DentistResponseDto(
+                dentist.getId(),
+                dentist.getRegistrationNumber(),
+                dentist.getLastName(),
+                dentist.getFirstName()
+        );
     }
 }
